I really enjoy pain. #8

+++ src/test/lib/optimizer/join_ordering/enumerate_ccp_test.cpp
@@ -10,8 +10,8 @@ template <typename T>
 bool equals(const std::pair<boost::dynamic_bitset<>, boost::dynamic_bitset<>>& lhs, const std::pair<T, T>& rhs) {
   Assert(lhs.first.size() == lhs.second.size() && lhs.first.size() <= sizeof(unsigned long) * 8,  // NOLINT
          "Bitset has too many bits for comparison.");
-  return lhs.first.to_ulong() == static_cast<unsigned long>(rhs.first) &&
-         lhs.second.to_ulong() == static_cast<unsigned long>(rhs.second);
+  return lhs.first.to_ulong() == static_cast<size_t>(rhs.first) &&
+         lhs.second.to_ulong() == static_cast<size_t>(rhs.second);
 }
 }  // namespace
